moving poller to pkg/controller for reuse elsewhere
